Actividad deudas y ReciclyeAdapterDebt añadido. Se ha cambiado el grid del main con las personas. Todo corregido para que funcione.

+++ app/src/main/java/com/example/moneyblacklist/Model/Debt.kt
@@ -1,3 +1,3 @@
 package com.example.moneyblacklist.Model
 
-class Debt (val concept: String, val amount: Float, val paid: Boolean)
+class Debt (val concept: String, val amount: Double, val paid: Boolean)
